Put resources back into namespaces
