+++ app/resources/api/v1/member_resource.rb
@@ -1,0 +1,19 @@
+module Api
+  module V1
+    class MemberResource < BaseResource
+      immutable
+
+      has_many :gardens
+      has_many :plantings
+      has_many :harvests
+      has_many :seeds
+
+      has_many :photos
+
+      attribute :login_name
+      attribute :slug
+
+      filters :login_name, :slug
+end
+end
+end
