--- app/resources/garden_resource.rb
@@ -1,9 +1,0 @@
-class GardenResource < BaseResource
-  immutable
-
-  has_one :owner, class_name: 'Member'
-  has_many :plantings
-  has_many :photos
-
-  attribute :name
-end
