--- app/resources/harvest_resource.rb
@@ -1,15 +1,0 @@
-class HarvestResource < BaseResource
-  immutable
-
-  has_one :crop
-  has_one :planting
-  has_one :owner, class_name: 'Member'
-  has_many :photos
-
-  attribute :harvested_at
-  attribute :description
-  attribute :unit
-  attribute :weight_quantity
-  attribute :weight_unit
-  attribute :si_weight
-end
